Made the GUI dynamic
I changed everything to the pack system instead of using the grid.

Now there are a lot of frames (right now they are colorful, so I can see them, I am going to change that later on). In these frames, all of the "items" are being packed and aligned to either top, bottom, right or left. Depending on how they are supposed to sit next to each other.