Add files via upload
Add the new LinearRegression in Algorithms Folder. This LinearRegression works already, but it doesn't include the GUI. In addition, the homeName and guestName are also as Input, because I need to check, if the algorithm is right. So when you link GUI with it, it maybe change the variable of homeName and guestName firstly.